code for saving message_state into database moved inside rule engine
