+++ kalbur/rule_engine/save_ms.h
@@ -1,0 +1,6 @@
+#ifndef SAVE_MS_H
+#define SAVE_MS_H
+
+int save_msg(sqlite3 *db, hashtable_t *hash_table, struct message_state *ms);
+
+#endif // SAVE_MS_H
